Packages in an R-universe

List all packages available in an R-universe.

runi_packages(universe, package = NULL, version = NULL)

Arguments

universe
A length one character. User or organization name of the the R-universe owner.

package
either NULL, or a length one character. Package name (see details).

version
either NULL, or a character vector. Package version (see details).

Value

a tibble if both package and version are not NULL, a character vector otherwise.

Details

This function gives access to three types of information, depending on how many arguments are different from NULL:

runi_packages(universe) lists all packages in the R-universe universe.

runi_packages(universe, package), where package is a string, lists all available version for package package in this universe.

runi_packages(universe, package, version) lists all available builds for package package of version version (specified as a string).

The argument version is ignored if package is NULL.

Author

Valerio Gherardi

Examples

( packages <- runi_packages("vgherard") )

#> [1] "fcci"     "gsample"  "kgrams"   "r2r"      "runiv"    "sbo"      "scribblr"

( versions <- runi_packages("vgherard", packages[[1]]) )

#> [1] "1.0.0"

( versions <- runi_packages("vgherard", packages[[1]], versions[[1]]) )

#> # A tibble: 5 × 32
#>   `_id`  Package Type   Title Version `Authors@R` Maintainer Description License
#>   <chr>  <chr>   <chr>  <chr> <chr>   <chr>       <chr>      <chr>       <chr>  
#> 1 60e02… fcci    Packa… Feld… 1.0.0   "person(gi… Valerio G… "Provides … GPL (>…
#> 2 60e02… fcci    Packa… Feld… 1.0.0   "person(gi… Valerio G… "Provides … GPL (>…
#> 3 60e02… fcci    Packa… Feld… 1.0.0   "person(gi… Valerio G… "Provides … GPL (>…
#> 4 60e02… fcci    Packa… Feld… 1.0.0   "person(gi… Valerio G… "Provides … GPL (>…
#> 5 60e02… fcci    Packa… Feld… 1.0.0   "person(gi… Valerio G… "Provides … GPL (>…
#> # … with 23 more variables: Encoding <chr>, RoxygenNote <chr>,
#> #   SystemRequirements <chr>, URL <chr>, BugReports <chr>,
#> #   Config/testthat/edition <chr>, Repository <chr>, RemoteUrl <chr>,
#> #   RemoteRef <chr>, RemoteSha <chr>, NeedsCompilation <chr>, Packaged <chr>,
#> #   Author <chr>, Built <df[,4]>, _user <chr>, _type <chr>, _file <chr>,
#> #   _published <chr>, _builder <df[,11]>, MD5sum <chr>, _hard_deps <list>,
#> #   _soft_deps <list>, Archs <chr>